feat: ネットバトル開始機能を追加し、オンラインチェックを実装
